Add files via upload
The aim of this project is to develop a mobile application that provides previous year question papers and their answers for all departments and subjects in a B.Tech college. The app aims to help students prepare for exams by providing them with a comprehensive repository of previous year papers, thereby reducing the time and effort required to search for these resources.